+++ Frontend/src/pages/User.jsx
@@ -102,15 +102,17 @@ const User = ({ self }) => {
 
   const handleDelete = () => {
     (async () => {
-      alert("Are you sure you want to delete this user?");
-      const accessToken = await getAccessTokenSilently();
-      axiosInstance
-      .delete(routes.profiles.deleteProfile(id), {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      })
-      .then(() => navigate("/users"));
+      const action = window.confirm("Are you sure you want to delete this user?");
+      if (action) {
+        const accessToken = await getAccessTokenSilently();
+        axiosInstance
+          .delete(routes.profiles.deleteProfile(id), {
+            headers: {
+              Authorization: `Bearer ${accessToken}`,
+            },
+          })
+          .then(() => navigate("/users"));
+        }
     })();
   };
 
@@ -129,15 +131,17 @@ const User = ({ self }) => {
 
   const handleCheckOut = () => {
     (async () => {
-      alert("Are you sure you want to checkout?");
-      const accessToken = await getAccessTokenSilently();
-      axiosInstance
-        .put(routes.bookings.checkOut(id), {ShelterId: userData.currentShelter?.shelterId}, {
-          headers: {
-            Authorization: `Bearer ${accessToken}`,
-          },
-        })
-        .then(() => { getUser(); getBookingHistory(); });
+      let action = window.confirm("Are you sure you want to checkout?");
+      if (action) {
+        const accessToken = await getAccessTokenSilently();
+        axiosInstance
+          .put(routes.bookings.checkOut(id), {ShelterId: userData.currentShelter?.shelterId}, {
+            headers: {
+              Authorization: `Bearer ${accessToken}`,
+            },
+          })
+          .then(() => { getUser(); getBookingHistory(); });
+        }
     })();
   };
 
@@ -171,14 +175,16 @@ const User = ({ self }) => {
           <Button onClick={() => setOpenedModal(true)}>
             <MdEdit /> Edit
           </Button>
-          <AdminOnly>
-            <Button
-              className="delete-button"
-              onClick={() => handleDelete()}
-            >
-              <MdDelete /> Delete
-            </Button>
-          </AdminOnly>
+          {userData.currentShelter == null &&
+            <AdminOnly>
+              <Button
+                className="delete-button"
+                onClick={() => handleDelete()}
+              >
+                <MdDelete /> Delete
+              </Button>
+            </AdminOnly>
+          }
         </div>
       </div>
       <div className="flex flex-col gap-10">
